Refactor product data structure and remove mock products data
