+++ src/app/products/page.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from 'react';
 import Image from 'next/image';
-import { productTypes } from '../../constants/data/mock-products-data';
+import { productTypes } from '../../constants/data/products-data';
 
 const productImage =
   'https://cdn.prod.website-files.com/658e0f2123ee3a398167d04f/67feb3ac5c609454b94e8066_E1.avif';
