+++ src/constants/data/base-cabinets-data.ts
@@ -1,0 +1,214 @@
+import { ProductType } from './products-data';
+
+export const baseCabinets: ProductType[] = [
+  {
+    name: 'Base Cabinets',
+    description: 'Floor-standing cabinets that provide the foundation for your kitchen',
+    products: [
+      {
+        name: 'Base Cabinet',
+        description: '1 drawer and 1 door',
+        features: ['1 Drawer', '1 Door', '1 Adjustable Shelf'],
+        image: 'BaseCabinets/1-1.jpg',
+        variants: [
+          {
+            code: 'B09',
+            dimensions: '9W x 34-1/2H x 24D',
+            weight: 42,
+          },
+          {
+            code: 'B12',
+            dimensions: '12W x 34-1/2H x 24D',
+            weight: 45,
+          },
+          {
+            code: 'B15',
+            dimensions: '15W x 34-1/2H x 24D',
+            weight: 50,
+          },
+          {
+            code: 'B18',
+            dimensions: '18W x 34-1/2H x 24D',
+            weight: 53,
+          },
+          {
+            code: 'B21',
+            dimensions: '21W x 34-1/2H x 24D',
+            weight: 57,
+          },
+        ],
+        availableStyles: [
+          'WhiteShaker',
+          'Dove',
+          'Charcoal',
+          'Sage',
+          'Pebble',
+          'Pure',
+          'Naval',
+          'ButterScotch',
+          'JavaCoffee',
+          'CastleGrey',
+          'Greige',
+          'Espresso',
+          'MochaGlazed',
+          'Hazel',
+          'CremeGlazed',
+          'Mahogany',
+        ],
+      },
+      {
+        name: 'Base Cabinet',
+        description: '1 drawer and 2 doors',
+        features: ['1 Drawer', '1 Door', '1 Adjustable Shelf'],
+        image: 'BaseCabinets/1-2.jpg',
+        variants: [
+          {
+            code: 'B24',
+            dimensions: '24W x 34-1/2H x 24D',
+            weight: 63,
+          },
+          {
+            code: 'B27',
+            dimensions: '27W x 34-1/2H x 24D',
+            weight: 68,
+          },
+        ],
+        availableStyles: [
+          'WhiteShaker',
+          'Dove',
+          'Charcoal',
+          'Sage',
+          'Pebble',
+          'Pure',
+          'Naval',
+          'ButterScotch',
+          'JavaCoffee',
+          'CastleGrey',
+          'Greige',
+          'Espresso',
+          'MochaGlazed',
+          'Hazel',
+          'CremeGlazed',
+          'Mahogany',
+        ],
+      },
+      {
+        name: 'Base Cabinet',
+        description: '2 drawer2 and 2 doors',
+        features: ['2 Drawer', '2 Door', '1 Adjustable Shelf'],
+        image: 'BaseCabinets/1-3.jpg',
+        variants: [
+          {
+            code: 'B30',
+            dimensions: '30W x 34-1/2H x 24D',
+            weight: 81,
+          },
+          {
+            code: 'B33',
+            dimensions: '33W x 34-1/2H x 24D',
+            weight: 85,
+          },
+          {
+            code: 'B36',
+            dimensions: '36W x 34-1/2H x 24D',
+            weight: 89,
+          },
+          {
+            code: 'B42',
+            dimensions: '42W x 34-1/2H x 24D',
+            weight: 94,
+          },
+        ],
+        availableStyles: [
+          'WhiteShaker',
+          'Dove',
+          'Charcoal',
+          'Sage',
+          'Pebble',
+          'Pure',
+          'Naval',
+          'ButterScotch',
+          'JavaCoffee',
+          'CastleGrey',
+          'Greige',
+          'Espresso',
+          'MochaGlazed',
+          'Hazel',
+          'CremeGlazed',
+          'Mahogany',
+        ],
+      },
+      {
+        name: 'Base Cabinet',
+        description: '3 drawers',
+        features: ['1 Small Drawer', '2 Large Drawers'],
+        image: 'BaseCabinets/1-4.jpg',
+        variants: [
+          {
+            code: 'DB12-3',
+            dimensions: '12W x 34-1/2H x 24D',
+            weight: 68,
+          },
+          {
+            code: 'DB15-3',
+            dimensions: '15W x 34-1/2H x 24D',
+            weight: 76,
+          },
+          {
+            code: 'DB18-3',
+            dimensions: '18W x 34-1/2H x 24D',
+            weight: 81,
+          },
+          {
+            code: 'DB21-3',
+            dimensions: '21W x 34-1/2H x 24D',
+            weight: 86,
+          },
+          {
+            code: 'DB24-3',
+            dimensions: '24W x 34-1/2H x 24D',
+            weight: 93,
+          },
+          {
+            code: 'DB30-3',
+            dimensions: '30W x 34-1/2H x 24D',
+            weight: 103,
+          },
+          {
+            code: 'DB33-3',
+            dimensions: '33W x 34-1/2H x 24D',
+            weight: 107,
+          },
+          {
+            code: 'DB36-3',
+            dimensions: '36W x 34-1/2H x 24D',
+            weight: 111,
+          },
+          {
+            code: 'DB42-3',
+            dimensions: '42W x 34-1/2H x 24D',
+            weight: 120,
+          },
+        ],
+        availableStyles: [
+          'WhiteShaker',
+          'Dove',
+          'Charcoal',
+          'Sage',
+          'Pebble',
+          'Pure',
+          'Naval',
+          'ButterScotch',
+          'JavaCoffee',
+          'CastleGrey',
+          'Greige',
+          'Espresso',
+          'MochaGlazed',
+          'Hazel',
+          'CremeGlazed',
+          'Mahogany',
+        ],
+      },
+    ],
+  },
+];
